Rename camel case
